+++ src/index.d.ts
@@ -1,0 +1,29 @@
+import Game from "./Game";
+import GameManager from "./GameManager";
+import GameConfig, { TableType } from "./GameConfig";
+
+import Human, { HumanId, SpouseId, Skills, Gender } from "./classes/entities/Human";
+import City from "./classes/City";
+import Agent, { AgentId, AgentType, AgentTypes } from "./classes/Agent";
+import Vector2 from "./classes/Vector2";
+
+import Database, { DatabaseResult } from "./databases/Database";
+import InMemoryDatabase from "./databases/InMemoryDatabase";
+
+export {
+    // Main types
+    Game,
+    GameManager,
+    GameConfig, TableType,
+
+    // Main classes
+    Human, HumanId, SpouseId, Skills, Gender,
+    City,
+    Agent, AgentId, AgentType, AgentTypes,
+    Vector2,
+
+    // Databases
+    Database,
+    DatabaseResult,
+    InMemoryDatabase
+};